+++ seantis/people/browser/export_form.py
@@ -1,6 +1,7 @@
 import codecs
 
 from copy import copy
+from cStringIO import StringIO as BytesIO
 from importlib import import_module
 
 from five import grok
@@ -10,6 +11,7 @@ from z3c.form import field
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from z3c.form.browser.radio import RadioFieldWidget
 from plone.directives import form
+from tablib import formats, packages
 
 from seantis.people import _
 from seantis.people.interfaces import IList
@@ -117,6 +119,30 @@ class ExportForm(BaseForm):
         yield dict(name='export', title=_(u"Export"), css_class='context')
         yield dict(name='cancel', title=_(u"Cancel"))
 
+    def export_as(self, dataset, format):
+        if format != 'xlsx':
+            return getattr(dataset, format)
+
+        # NOTE expect breakage here, tablib packages its own openpyxl module.
+        # (It's the package maintainers "style" to do this). If something stops
+        # working here, check back with tablib, if they updated the openpyxl
+        # module and have a look at _xlsx.py in tablib.
+        wb = packages.openpyxl.workbook.Workbook()
+        ws = wb.worksheets[0]
+        ws.title = dataset.title if dataset.title else 'Tablib Dataset'
+
+        # activate the autofilter for all columns
+        ws.auto_filter = ':'.join(cell.address for cell in (
+            ws.cell(row=0, column=0),
+            ws.cell(row=0, column=dataset.width)
+        ))
+
+        formats._xlsx.dset_sheet(dataset, ws)
+
+        stream = BytesIO()
+        wb.save(stream)
+        return stream.getvalue()
+
     def handle_export(self):
         try:
             export_fields = self.parameters.get('export_fields')
@@ -137,7 +163,7 @@ class ExportForm(BaseForm):
             filename = '%s.%s' % (self.context.title, format)
             filename = codecs.utf_8_encode('filename="%s"' % filename)[0]
 
-            output = getattr(dataset, format)
+            output = self.export_as(dataset, format)
 
             RESPONSE = self.request.RESPONSE
             RESPONSE.setHeader("Content-disposition", filename)